+++ src/openrct2-ui/drawing/engines/opengl/OpenGLDrawingEngine.cpp
@@ -32,7 +32,7 @@
 #    include <openrct2/drawing/IDrawingContext.h>
 #    include <openrct2/drawing/IDrawingEngine.h>
 #    include <openrct2/drawing/LightFX.h>
-#    include <openrct2/drawing/Rain.h>
+#    include <openrct2/drawing/Weather.h>
 #    include <openrct2/interface/Screenshot.h>
 #    include <openrct2/ui/UiContext.h>
 #    include <unordered_map>
