+++ src/openrct2/drawing/Weather.h
