+++ src/openrct2/libopenrct2.vcxproj
@@ -189,7 +189,7 @@
     <ClInclude Include="drawing\ImageImporter.h" />
     <ClInclude Include="drawing\LightFX.h" />
     <ClInclude Include="drawing\NewDrawing.h" />
-    <ClInclude Include="drawing\Rain.h" />
+    <ClInclude Include="drawing\Weather.h" />
     <ClInclude Include="drawing\Text.h" />
     <ClInclude Include="drawing\TTF.h" />
     <ClInclude Include="drawing\X8DrawingEngine.h" />
@@ -510,7 +510,7 @@
     <ClCompile Include="drawing\LightFX.cpp" />
     <ClCompile Include="drawing\Line.cpp" />
     <ClCompile Include="drawing\NewDrawing.cpp" />
-    <ClCompile Include="drawing\Rain.cpp" />
+    <ClCompile Include="drawing\Weather.cpp" />
     <ClCompile Include="drawing\Rect.cpp" />
     <ClCompile Include="drawing\ScrollingText.cpp" />
     <ClCompile Include="drawing\SSE41Drawing.cpp" />
